# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Kino API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/cinema/address_name": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение всех названий и адресов кинотеатров",
                "responses": {
                    "200": {
                        "description": "Список всех названий и адресов кинотеатров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.CinemaAddressName"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/categories": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение всех категорий кинотеатра",
                "responses": {
                    "200": {
                        "description": "Список всех категорий кинотеатра",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.CinemaCategory"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/conditions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение всех состояний кинотеатра",
                "responses": {
                    "200": {
                        "description": "Список всех состояний кинотеатра",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.CinemaCondition"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/hall/types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение всех типов зала кинотеатра",
                "responses": {
                    "200": {
                        "description": "Список всех типов зала кинотеатра",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.CinemaHallType"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/halls/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение всех залов кинотеатра по его ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кинотеатра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Залы кинотеатра успешно получены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.GetCinemaHall"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "404": {
                        "description": "Зал не найден",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/id/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cinema"
                ],
                "summary": "Получение полной информации о кинотеатре по его ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кинотеатра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Полная информация о кинотеатре получена",
                        "schema": {
                            "$ref": "#/definitions/entities.GetCinema"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "404": {
                        "description": "Зал не найден",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/cinema/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Удаление кинотеатра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кинотеатра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Кинотеатр успешно удален",
                        "schema": {
                            "$ref": "#/definitions/entities.ID"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "404": {
                        "description": "Кинотеатр не найден",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.CinemaAddressName": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.CinemaCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.CinemaCondition": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.CinemaHallType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entities.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "entities.GetCinema": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                }
            }
        },
        "entities.GetCinemaHall": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "entities.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}